+++ src/views/notifications/PipelinesMonitor.ts
@@ -1,12 +1,18 @@
 import { commands, window } from 'vscode';
 
+import { notificationBannerClickedEvent, notificationChangeEvent } from '../../analytics';
 import { clientForSite } from '../../bitbucket/bbUtils';
 import { WorkspaceRepo } from '../../bitbucket/model';
+import { configuration } from '../../config/configuration';
 import { Commands } from '../../constants';
 import { Container } from '../../container';
 import { Pipeline, PipelineTarget } from '../../pipelines/model';
 import { BitbucketActivityMonitor } from '../BitbucketActivityMonitor';
 import { descriptionForState, generatePipelineTitle, shouldDisplay } from '../pipelines/Helpers';
+import { NotificationSurface } from './notificationManager';
+import { NotificationSource } from './notificationSources';
+
+const NotificationBannerId = 'atlascode.bitbucket.pipelinesNotification';
 
 export class PipelinesMonitor implements BitbucketActivityMonitor {
     private _previousResults: Record<string, Pipeline[]> = {};
@@ -17,6 +23,7 @@ export class PipelinesMonitor implements BitbucketActivityMonitor {
         if (!Container.config.bitbucket.pipelines.monitorEnabled) {
             return;
         }
+
         for (let i = 0; i < this._repositories.length; i++) {
             const wsRepo = this._repositories[i];
             const previousResults = this._previousResults[wsRepo.rootUri];
@@ -25,29 +32,51 @@ export class PipelinesMonitor implements BitbucketActivityMonitor {
             if (!site) {
                 return;
             }
-            const bbApi = await clientForSite(site);
 
+            const bbApi = await clientForSite(site);
             if (!bbApi.pipelines) {
                 return; //Bitbucket Server instances will not have pipelines
             }
 
-            bbApi.pipelines.getRecentActivity(site).then((newResults) => {
-                let diffs = this.diffResults(previousResults, newResults);
-                diffs = diffs.filter((p) => this.shouldDisplayTarget(p.target));
+            const newResults = await bbApi.pipelines.getRecentActivity(site);
+            const diffs = this.diffResults(previousResults, newResults).filter((p) =>
+                this.shouldDisplayTarget(p.target),
+            );
+
+            if (diffs.length > 0) {
                 const buttonText = diffs.length === 1 ? 'View' : 'View Pipeline Explorer';
-                if (diffs.length > 0) {
-                    window.showInformationMessage(this.composeMessage(diffs), buttonText).then((selection) => {
-                        if (selection) {
-                            if (diffs.length === 1) {
-                                commands.executeCommand(Commands.ShowPipeline, diffs[0]);
-                            } else {
-                                commands.executeCommand('workbench.view.extension.atlascode-drawer');
-                            }
-                        }
-                    });
+                const neverShow = "Don't show again";
+
+                notificationChangeEvent(
+                    NotificationSource.BitbucketPipeline,
+                    undefined,
+                    NotificationSurface.Banner,
+                    1,
+                ).then((event) => {
+                    Container.analyticsClient.sendTrackEvent(event);
+                });
+
+                const selection = await window.showInformationMessage(
+                    this.composeMessage(diffs),
+                    buttonText,
+                    neverShow,
+                );
+                if (selection === buttonText) {
+                    if (diffs.length === 1) {
+                        commands.executeCommand(Commands.ShowPipeline, diffs[0]);
+                    } else {
+                        commands.executeCommand('workbench.view.extension.atlascode-drawer');
+                    }
+                } else if (selection === neverShow) {
+                    configuration.updateEffective('bitbucket.pipelines.monitorEnabled', false, null, true);
                 }
+
+                notificationBannerClickedEvent(NotificationBannerId, selection ?? '').then((event) => {
+                    Container.analyticsClient.sendUIEvent(event);
+                });
+            } else {
                 this._previousResults[wsRepo.rootUri] = newResults;
-            });
+            }
         }
     }
 
@@ -56,21 +85,20 @@ export class PipelinesMonitor implements BitbucketActivityMonitor {
         return !target.ref_name || shouldDisplay(target);
     }
 
+    // TODO check this function: it doesn't seem correct, it skips results
     private diffResults(oldResults: Pipeline[], newResults: Pipeline[]): Pipeline[] {
         if (!oldResults) {
             return [];
         }
+
         const changes: Pipeline[] = [];
-        const previousLength = oldResults.length;
-        const newLength = newResults.length;
+
         let i = 0;
         let j = 0;
-        while (true) {
-            if (i === previousLength || j === newLength) {
-                return changes;
-            }
+        while (i < oldResults.length && j < newResults.length) {
             const oldItem = oldResults[i];
             const newItem = newResults[j];
+
             if (oldItem.build_number === newItem.build_number) {
                 if (oldItem.state!.name !== newItem.state!.name) {
                     changes.push(newItem);
@@ -82,6 +110,8 @@ export class PipelinesMonitor implements BitbucketActivityMonitor {
                 j++;
             }
         }
+
+        return changes;
     }
 
     private composeMessage(newResults: Pipeline[]): string {